+++ src/client/components/info_btn/info_btn.js
@@ -19,6 +19,8 @@ Template.info_btn.onCreated( function(){
     self.PCK = {
         // have a deep copy of provided data context
         dataContext: new ReactiveVar( null ),
+        // whether the 'info' button is hidden when disabled
+        hideWhenDisabled: new ReactiveVar( Tabular.configure().hideDisabled ),
         // manage the UI
         enabled: new ReactiveVar( true ),
         title: new ReactiveVar( '' )
@@ -35,6 +37,9 @@ Template.info_btn.onCreated( function(){
                 .then(( res ) => { self.PCK.enabled.set( res ); })
             dataContext.table.opt( 'infoButtonTitle', pwixI18n.label( I18N, 'info.btn_title', dataContext.item._id ), dataContext.item )
                 .then(( res ) => { self.PCK.title.set( res ); });
+            // hide when disabled ?
+            this.opt( 'hideDisabled', Tabular.configure().hideDisabled )
+                .then(( res ) => { self.PCK.hideWhenDisabled.set( res ); });
             // run only once
             comp.stop();
         }
@@ -59,7 +64,7 @@ Template.info_btn.helpers({
 
     // whether to show the disabled button ?
     show(){
-        return Template.instance().PCK.enabled.get() || !Tabular.configure().hideDisabled;
+        return Template.instance().PCK.enabled.get() || !Template.instance().PCK.hideWhenDisabled.get();
     }
 });